requirements.txt
streamlit
langchain
langchain-community
python-dotenv
pypdf
langchain_chroma
langchain_huggingface
langchain-core
sentence-transformers
huggingface_hub[cli]
unstructured
unstructured[pdf]
wikipedia
langchain
langchain-community
python-dotenv
langchain-huggingface
langchain-groq
langchain_chroma
pypdf
sentence-transformers
grandalf